added copyright header and renamed markupSetting to markupSettingPanel
