+++ packages/markup-frontstage/src/components/toolbar/MarkupToolWidget.module.scss
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
 .verticalToolbar {
   margin-top: 11px;
 }
