+++ packages/markup-frontstage/src/markup-frontstage-react.ts
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 export * from "./components/ui/MarkupFrontstageProvider";
-export * from "./components/toolbar/MarkupSettings";
+export * from "./components/toolbar/MarkupSettingsPanel";
 export * from "./util/MarkupTypes";
