+++ packages/markup-frontstage/src/tests/components/toolbar/MarkupSettings.test.tsx
@@ -1,8 +1,12 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 import { Point } from "@bentley/ui-core";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import React from "react";
 
-import MarkupSettingsPanel from "../../../components/toolbar/MarkupSettings";
+import MarkupSettingsPanel from "../../../components/toolbar/MarkupSettingsPanel";
 import { MarkupFrontstage } from "../../../MarkupFrontstage";
 import { I18N } from "@bentley/imodeljs-i18n";
 
@@ -21,7 +25,7 @@ jest.mock("@bentley/imodeljs-i18n", () => ({
 }));
 jest.mock("../../../MarkupFrontstage");
 
-describe("MarkupSettings tool testing", () => {
+describe("MarkupSettingsPanel tool testing", () => {
   it("should render", () => {
     //arrange
     MarkupFrontstage.initialize(new I18N());
