+++ packages/markup-frontstage/src/tests/components/toolbar/MarkupToolWidget.test.tsx
@@ -1,5 +1,7 @@
-// Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 import { I18N } from "@bentley/imodeljs-i18n";
 import { UiFramework } from "@bentley/ui-framework";
 import { render } from "@testing-library/react";
