+++ packages/markup-frontstage/src/tests/mocks/i18n.js
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 import i18n from "i18next";
 import { I18N } from "@bentley/imodeljs-i18n";
