+++ packages/markup-frontstage/src/util/MarkupTypes.ts
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 import { MarkupUiEvent } from "./MarkupUiEvents";
 import { SavedViewData } from "./SavedViewTypes";
